subvortex/auto_upgrader/requirements.txt
bittensor==9.0.4
bittensor-wallet==3.0.4
python-dotenv==1.0.1
requests==2.32.3